Move validation of index_order to ParticleIndex, begin using ParticleIndexInfo class
